Cleanup directus_user_avatar interface
