+++ app/core/interfaces/_internals/user_avatar/interface.js
@@ -1,0 +1,14 @@
+define(['app', 'core/UIView'], function (app, UIView) {
+  'use strict';
+
+  return UIView.extend({
+    tagName: 'fieldset',
+    initialize: function (options) {
+      var user = app.users.get(options.value);
+      if (user) {
+        var avatar = user.getAvatar();
+        this.$el.append('<img src="' + avatar + '" class="avatar" />');
+      }
+    }
+  });
+});
